refactor: #1718 컴포넌트 이름 수정 및 아이콘 인식 범위 증가
