+++ src/feature/board/component/AccordianBoards/AccordianBoards.module.css
@@ -17,6 +17,7 @@
 .arrow {
   transition: transform 0.2s ease;
   cursor: pointer;
+  padding-left: 1rem;
 }
 
 .rotated {
